refactor: 별점 UI 업데이트와 저장 로직 분리
