+++ src/constants/rating.ts
@@ -1,0 +1,9 @@
+export const RATING_TEXTS = [
+  "최악이에요",
+  "별로에요",
+  "보통이에요",
+  "재밌어요",
+  "명작이에요",
+];
+
+export const RATING_SCORES = ["2", "4", "6", "8", "10"];
