+++ src/services/storage.ts
@@ -1,0 +1,3 @@
+export const setLocalStorage = (key: string, value: string) => {
+  window.localStorage.setItem(key, value);
+};
